backend/requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.0.1
Flask-Mail==0.9.1
python-dotenv==0.19.0
SQLAlchemy==1.4.22
psycopg2-binary==2.9.1
gunicorn==20.1.0
openai==0.10.2
requests==2.25.1
python-jose==3.3.0
passlib==1.7.4
Flask-JWT-Extended==4.3.1
pandas==1.3.3
numpy==1.21.2
matplotlib==3.4.3
seaborn==0.11.2
beautifulsoup4==4.9.3
lxml==4.6.3
celery==5.1.2
redis==3.5.3
flower==1.0.0
PyJWT==2.1.0
cryptography==3.4.7
email-validator==1.1.3